requirements.txt
requests~=2.25.1
urllib3~=1.26.4
pytest~=6.2.4
PyJWT~=2.1.0
YURL~=1.0.0
